refactor(core): one externals list, and a CDN probe that stops waiting on a server (v2 §1.6 review)
Four findings from the quality review of this PR. No behavior change anywhere:
`dist/engine/index.global.js`, `dist/index.js` and `dist/index.cjs` are all
md5-identical before and after, and the size rows are unmoved (17 492 / 18 098
/ 22 857).

1. `tsup.config.ts` stated the five externals TWICE, once per config item, plus
   five more output keys. That is the exact drift this file exists to prevent:
   a sixth external added to the main item alone would be BUNDLED into the IIFE
   — not emitted as the throwing `__require()` shim the item comment warns
   about, just silently inlined, where the `require(` guard cannot see it and
   the `core:engine:iife` row absorbs the bytes without a word. Hoisted to one
   `EXTERNAL` and one `SHARED` output policy; the items now differ only in
   entry, format, platform and clean, which is what a reader wants to compare.

2. `probe:cdn` ran behind `next dev` and the page probe, though it reads a file
   and runs it in a `vm` — no port, no page, no server. That inverted the
   failure mode: a release that breaks the app's boot is exactly when you want
   to know whether the tarball is intact, and behind the boot you learned
   nothing. It runs first now. Full `run-smoke.sh` verified green in the new
   order.

3. The manifest describe was the only non-`skipIf(!distExists())` block in a
   dist-tier test file — a different tier wearing the wrong file. Moved to a new
   `manifest.test.ts`, the home `vue`, `svelte` and `analytics` each have and
   core lacked; core's manifest facts had been accumulating in whichever file
   the feature that added them lived in.

4. `realm() as unknown as { TourKit: TourKitGlobal }` was a double cast at the
   call site. The helper returns `{ ctx, TourKit }` now, so the one unavoidable
   cast (`vm.Context` is `object`) sits in one place instead of three. While
   there, the target-less tour moved up beside the targeted one as a named
   fixture — the missing targets are the point of the realm case, not an
   inline detail.

Both red-first guards re-verified against the `platform`-less build after the
restructure: the `process.env` grep and the realm's `ReferenceError: process is
not defined` still fail, and nothing else does.
