+++ apps/smoke/scripts/run-smoke.sh
@@ -17,6 +17,15 @@ cleanup() {
 }
 trap cleanup EXIT
 
+# v2 §1.6 — the CDN build, checked FIRST. It reads a file out of the installed
+# tarball and runs it in a `vm`: no port, no page, no dev server. Gating it
+# behind `next dev` would invert the failure mode — a release that breaks the
+# app's boot is exactly when you want to know whether the tarball is intact, and
+# behind the boot you would learn nothing. It is also the only step in this
+# script that executes the published JavaScript; `probe` below is curl.
+log "probing the CDN build"
+pnpm probe:cdn
+
 log "booting next dev on :3100"
 pnpm dev > /tmp/tour-kit-smoke-dev.log 2>&1 &
 DEV_PID=$!
@@ -38,12 +47,6 @@ done
 log "probing http://localhost:3100/"
 if pnpm probe; then
   log "OK — smoke page rendered with data-smoke-ok marker"
-  # v2 §1.6 — the CDN build. `probe` above is curl, which executes no
-  # JavaScript, so this is the only thing here that actually RUNS the tarball.
-  # `set -e` is in force inside this branch (it is suspended only for the
-  # command in the `if` condition), so a non-zero exit aborts the script.
-  log "probing the CDN build"
-  pnpm probe:cdn
   exit 0
 else
   log "FAIL — smoke probe could not find data-smoke-ok marker"
